change to b16

+++ rp_handler.py
@@ -10,9 +10,9 @@ from runpod.serverless.utils.rp_download import file as rp_file
 from runpod.serverless.modules.rp_logger import RunPodLogger
 
 # --------------------------- КОНСТАНТЫ ----------------------------------- #
-MAX_SEED = np.iinfo(np.int32).max
+MAX_SEED = np.iinfo(np.int16).max
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
-DTYPE = torch.float16 if DEVICE == "cuda" else torch.float32
+DTYPE = torch.bfloat16 if DEVICE == "cuda" else torch.float32
 MAX_STEPS = 250
 
 TARGET_RES = 1024
